pyproject.toml
[project]
name = "pdf-epub-reader-with-gemini"
version = "0.1.0"
description = "AI-assisted PDF/EPUB reader with Gemini API integration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "PySide6>=6.7",
    "PyMuPDF>=1.24",
    "google-generativeai>=0.8",
    "qasync>=0.27",
    "python-dotenv>=1.0",
    "Pillow>=10.0",
    "platformdirs>=4.0",
    "markdown>=3.5",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pdf_epub_reader"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
